main.py
#!/usr/bin/env python3
"""
Agentes Autónomos y Sistemas "Agentivos"
Monitoreo de descargas y salud de sitios web
"""

import asyncio
import time
from datetime import datetime
from playwright.async_api import async_playwright
import discord
from discord.ext import tasks

# Configuración
URLS_MONITOREO = [
    "https://example.com",
    "https://tu-sitio.com"
]
DISCORD_WEBHOOK_URL = "tu-webhook-url"
INTERVALO_MINUTOS = 5

class AgenteMonitoreo:
    """Agente autónomo para monitoreo de sitios web"""

    def __init__(self):
        self.resultados = []
        self.esta_corriendo = False

    async def verificar_sitio(self, url: str) -> dict:
        """Verifica la salud de un sitio web usando Playwright"""
        resultado = {
            "url": url,
            "timestamp": datetime.now().isoformat(),
            "estado": "desconocido",
            "tiempo_carga": 0,
            "errores": []
        }

        try:
            async with async_playwright() as p:
                browser = await p.chromium.launch(headless=True)
                page = await browser.new_page()

                inicio = time.time()
                response = await page.goto(url, wait_until="networkidle", timeout=30000)
                tiempo_carga = time.time() - inicio

                resultado["tiempo_carga"] = round(tiempo_carga, 2)

                if response and response.status == 200:
                    resultado["estado"] = "operativo"
                    # Verificar elementos críticos
                    try:
                        await page.wait_for_selector("body", timeout=5000)
                    except Exception as e:
                        resultado["errores"].append(f"Elemento body no encontrado: {e}")
                else:
                    resultado["estado"] = "error"
                    resultado["errores"].append(f"Status: {response.status if response else 'No response'}")

                await browser.close()

        except Exception as e:
            resultado["estado"] = "fallo"
            resultado["errores"].append(str(e))

        return resultado

    async def autocorregir(self, url: str, error: str):
        """Intenta autocorregir problemas detectados"""
        print(f"[{datetime.now()}] Intentando autocorrección para {url}")
        # Aquí iría la lógica de autocorrección
        # Por ejemplo: reiniciar servicios, limpiar caché, etc.
        pass

    async def enviar_reporte_discord(self, resultados: list):
        """Envía reporte vía Discord webhook"""
        mensaje = f"**📊 Reporte de Monitoreo - {datetime.now().strftime('%Y-%m-%d %H:%M')}**\n\n"

        for r in resultados:
            emoji = "✅" if r["estado"] == "operativo" else "❌"
            mensaje += f"{emoji} **{r['url']}**\n"
            mensaje += f"   Estado: {r['estado']}\n"
            mensaje += f"   Tiempo de carga: {r['tiempo_carga']}s\n"
            if r["errores"]:
                mensaje += f"   Errores: {', '.join(r['errores'])}\n"
            mensaje += "\n"

        # Aquí implementarías el envío real al webhook de Discord
        print(mensaje)

    async def ciclo_monitoreo(self):
        """Ejecuta un ciclo completo de monitoreo"""
        print(f"\n[{datetime.now()}] Iniciando ciclo de monitoreo...")

        tareas = [self.verificar_sitio(url) for url in URLS_MONITOREO]
        resultados = await asyncio.gather(*tareas)
        self.resultados = resultados

        # Verificar si hay errores para autocorregir
        for r in resultados:
            if r["estado"] != "operativo":
                await self.autocorregir(r["url"], str(r["errores"]))

        await self.enviar_reporte_discord(resultados)
        print(f"[{datetime.now()}] Ciclo completado.\n")

    async def iniciar(self):
        """Inicia el agente de monitoreo en bucle infinito"""
        self.esta_corriendo = True
        print(f"🚀 Agente de Monitoreo iniciado - Intervalo: {INTERVALO_MINUTOS} minutos")

        while self.esta_corriendo:
            await self.ciclo_monitoreo()
            await asyncio.sleep(INTERVALO_MINUTOS * 60)

    def detener(self):
        """Detiene el agente"""
        self.esta_corriendo = False
        print("🛑 Agente detenido")


class OrquestadorAgentes:
    """Orquesta múltiples agentes distribuidos"""

    def __init__(self):
        self.agentes = {}

    def registrar_agente(self, nombre: str, agente: AgenteMonitoreo):
        """Registra un nuevo agente en el orquestador"""
        self.agentes[nombre] = agente
        print(f"✓ Agente '{nombre}' registrado")

    async def ejecutar_todos(self):
        """Ejecuta todos los agentes registrados"""
        if not self.agentes:
            print("No hay agentes registrados")
            return

        tareas = [agente.ciclo_monitoreo() for agente in self.agentes.values()]
        await asyncio.gather(*tareas)


if __name__ == "__main__":
    # Crear y ejecutar agente
    agente = AgenteMonitoreo()

    try:
        asyncio.run(agente.iniciar())
    except KeyboardInterrupt:
        agente.detener()
