requirements.txt
playwright>=1.40.0
discord.py>=2.3.0
asyncio-mqtt>=0.13.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
